requirements.txt
qrcode==8.0
pillow==11.1.0 